@@ -1,21 +1,22 @@
 {
-  "name": "kidsloop-live-webrtc-service",
-  "version": "1.2.0",
+  "name": "kidsloop-live-server",
+  "version": "1.3.1",
   "description": "",
-  "main": "index.js",
+  "main": "main.js",
   "scripts": {
-    "start": "ts-node src/main.ts",
-    "build": "webpack",
-    "test": "mocha -r ts-node/register tests/**/*[Tt]est.ts",
+    "start": "ts-node dist/index.js",
+    "dev": "ts-node-dev src/main.ts",
+    "build": "webpack --stats-modules-space 999",
+    "test": "",
     "coverage": "nyc -r lcov -e .ts -x \"*[Tt]est.ts\" npm run test"
   },
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@bitbucket.org/calmisland/kidsloop-live-webrtc-service.git"
+    "url": "git+ssh://git@bitbucket.org/calmisland/kidsloop-live-server.git"
   },
   "author": "",
   "license": "ISC",
-  "homepage": "https://bitbucket.org/calmisland/kidsloop-live-webrtc-service#readme",
+  "homepage": "https://bitbucket.org/calmisland/kidsloop-live-server#readme",
   "devDependencies": {
     "@types/axios": "^0.14.0",
     "@types/chai": "^4.2.11",
@@ -26,47 +27,49 @@
     "@types/newrelic": "^7.0.2",
     "@types/uuid": "^7.0.4",
     "@types/ws": "^7.2.5",
-    "@typescript-eslint/eslint-plugin": "^3.2.0",
-    "@typescript-eslint/parser": "^3.2.0",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "eslint": "^7.2.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-plugin-import": "^2.21.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.20.0",
-    "eslint-plugin-standard": "^4.0.1",
+    "eslint": "^7.32.0",
     "mocha": "^8.1.3",
     "nyc": "^15.1.0",
-    "ts-loader": "^7.0.5",
+    "ts-loader": "^9.2.6",
     "ts-mockito": "^2.6.1",
     "ts-node": "^8.10.2",
+    "ts-node-dev": "^1.1.8",
     "typescript": "^4.3.5",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11"
+    "webpack": "^5.58.2",
+    "webpack-cli": "^4.9.1"
   },
   "dependencies": {
     "@newrelic/apollo-server-plugin": "^1.0.2",
-    "@sentry/node": "^5.20.1",
     "@types/jsonwebtoken": "^8.5.0",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
     "apollo-server": "^2.14.3",
     "apollo-server-express": "^2.15.0",
     "aws-sdk": "^2.926.0",
-    "axios": "^0.21.1",
+    "axios": "^0.21.4",
     "cookie": "^0.4.1",
     "dotenv": "^10.0.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^5.1.0",
+    "eslint-plugin-react": "^7.26.1",
+    "eslint-plugin-standard": "^5.0.0",
     "express": "^4.17.1",
     "graphql": "^15.1.0",
+    "graphql-request": "^3.5.0",
     "graphql-scalars": "^1.8.0",
     "ioredis": "^4.17.3",
     "jsonwebtoken": "^8.5.1",
-    "kidsloop-token-validation": "bitbucket:calmisland/kidsloop-token-validation#v1.2.3",
+    "kidsloop-token-validation": "bitbucket:calmisland/kidsloop-token-validation",
     "logrocket": "^1.0.12",
     "newrelic": "^8.4.0",
     "pg": "^8.5.1",
     "typeorm": "^0.2.29",
     "uuid": "^8.1.0",
+    "webpack-node-externals": "^3.0.0",
     "ws": "^7.4.6"
   }
 }
